+++ ruoyi-vue-pro/yudao-module-hospital/src/main/java/cn/iocoder/yudao/module/hospital/service/medicine/MedicineServiceImpl.java
@@ -6,6 +6,7 @@ import cn.iocoder.yudao.module.hospital.controller.admin.medicine.vo.MedicinePag
 import cn.iocoder.yudao.module.hospital.controller.admin.medicine.vo.MedicineSaveReqVO;
 import cn.iocoder.yudao.module.hospital.dal.dataobject.MedicineDO;
 import cn.iocoder.yudao.module.hospital.dal.mysql.MedicineMapper;
+import cn.iocoder.yudao.module.hospital.framework.security.HospitalSecurityContext;
 import org.springframework.stereotype.Service;
 import javax.annotation.Resource;
 import java.util.Collection;
@@ -20,9 +21,12 @@ public class MedicineServiceImpl implements MedicineService {
 
     @Resource
     private MedicineMapper medicineMapper;
+    @Resource
+    private HospitalSecurityContext securityContext;
 
     @Override
     public Long createMedicine(MedicineSaveReqVO createReqVO) {
+        securityContext.requireAdmin();
         MedicineDO medicine = BeanUtils.toBean(createReqVO, MedicineDO.class);
         medicineMapper.insert(medicine);
         return medicine.getId();
@@ -30,6 +34,7 @@ public class MedicineServiceImpl implements MedicineService {
 
     @Override
     public void updateMedicine(MedicineSaveReqVO updateReqVO) {
+        securityContext.requireAdmin();
         validateMedicineExists(updateReqVO.getId());
         MedicineDO updateObj = BeanUtils.toBean(updateReqVO, MedicineDO.class);
         medicineMapper.updateById(updateObj);
@@ -37,6 +42,7 @@ public class MedicineServiceImpl implements MedicineService {
 
     @Override
     public void deleteMedicine(Long id) {
+        securityContext.requireAdmin();
         validateMedicineExists(id);
         medicineMapper.deleteById(id);
     }
